// src/app/models/user-session.model.ts
export class UserSession {
  id?: number;
  last_login?: string;
  is_superuser?: boolean;
  username?: string;
  first_name?: string;
  last_name?: string;
  email?: string;
  is_staff?: boolean;
  is_active?: boolean;
  date_joined?: string;
  groups?: number[];
  user_permissions?: number[];
  projects?: number[];
  access?: string;
  refresh?: string;
}